updates in csr conversion

+++ src/analytics/gunrock_bfs.hpp
@@ -34,7 +34,7 @@ using rship_predicate = std::function<bool(relationship&)>;
 /**
  * Typedef for a function that computes the weight of a relationship.
  */
-using rship_weight = std::function<float(relationship&)>;
+using rship_weight = std::function<double(relationship&)>;
 
 /**
  * Struct containing the results of SSSP
